Work: by using from java interface Serializable. Create binary serialization file(.ser). by using from jackson tools. Create JSON serialization file(.json)
Problems:
I was't figure out how to use jackson to serial Java Object. There are some problems with serial file(.json). It lose many parameters and attributions.

+++ src/main/java/LibDL/outdated/nn/layer/PoolingLayer.java
@@ -92,7 +92,7 @@ public class PoolingLayer extends DefaultLayer {
                     for (int k = 0; k < input.shape()[1]; k++) {
                         double item = input.getDouble(j, k, i);
                         int mem = maxMem.remove();
-                        int offsetInner = (int) (mem / kernelWidth) * inputWidth + mem % kernelWidth;
+                        int offsetInner = (mem / kernelWidth) * inputWidth + mem % kernelWidth;
                         int offsetOuter = j * inputWidth * stride + k * stride + inputWidth * inputHeight * i;
                         result.putScalarUnsafe(offsetInner + offsetOuter, item);
                     }
